Buffer??
Map generation is still too simple, just Perlin noise. However, I am currently trying to implement a buffer instead of the old scrolling camera. This will mean the grid is represented by a deque of deques, and whenever the player moves, for instance, right, tiles that are very far to the left are deleted and new tiles very far to the right are created. Currently glitchy but has potential!